requirements.txt
push2-python @ git+https://github.com/ffont/push2-python.git
python-osc>=1.8.0
python-rtmidi>=1.5.0
Pillow>=10.0.0
PyYAML>=6.0
python-dotenv>=1.0.0
numpy
